requirements.txt
fastapi
uvicorn
pymongo
python-multipart
python-dotenv 
